news-summarization-app/requirements.txt
langchain
newsapi-python
openai
chromadb
tiktoken
python-dotenv